+++ client/src/components/index.js
@@ -1,6 +1,5 @@
 export {Home} from './Home/Home';
 export {Profile} from './Profile/Profile';
 export {Page404} from './Page404/Page404';
-export {SimpleSlider} from './SimpleSlider';
+// export {SimpleSlider} from './SimpleSlider';
 export {Categories} from './Categories'
-export {CategoriesSecond} from './CategoriesSecond'
